HW3/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.3.0
beautifulsoup4==4.13.3
bs4==0.0.2
click==8.1.8
colorama==0.4.6
duckduckgo_search==7.5.2
fastapi==0.115.11
frozenlist==1.5.0
h11==0.14.0
idna==3.10
lxml==5.3.1
multidict==6.1.0
primp==0.14.0
propcache==0.3.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
sniffio==1.3.1
soupsieve==2.6
starlette==0.46.1
typing_extensions==4.12.2
uvicorn==0.34.0
yarl==1.18.3
